pyproject.toml
[project]
name = "ai-campus-admin-agent-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "openai-agents>=0.3.1",
    "pymongo>=4.15.1",
    "uvicorn>=0.36.0",
]
